% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProjectMouseTCEx.R
\docType{class}
\name{TrenaProjectMouseTCEx-class}
\alias{TrenaProjectMouseTCEx-class}
\alias{.TrenaProjectMouseTCEx}
\alias{TrenaProjectMouseTCEx}
\title{TrenaProjectMouseTCEx-class}
\usage{
TrenaProjectMouseTCEx(quiet = TRUE)
}
\value{
An object of the TrenaProjectMouseTCEx class
}
\description{
Expression, variant and covariate data for the genes of interest (perhaps unbounded) for pre-term birth studies
}
